naming changes / helper refactor --> devicedata
