+++ Bonsai.Tinkerforge/Bonsai.Tinkerforge.csproj
@@ -15,7 +15,6 @@
     <TargetFramework>net472</TargetFramework>
     <Features>strict</Features>
     <Version>0.1.0</Version>
-    <DebugType>embedded</DebugType>
   </PropertyGroup>
 
   <ItemGroup>
