+++ Bonsai.Tinkerforge/BrickletAmbientLightV3.cs
@@ -3,7 +3,6 @@ using System.ComponentModel;
 using System.Linq;
 using System.Reactive.Linq;
 using Tinkerforge;
-using System.Xml.Serialization;
 
 namespace Bonsai.Tinkerforge
 {
@@ -13,7 +12,7 @@ namespace Bonsai.Tinkerforge
     {
         [Description("Device data including address UID.")]
         [TypeConverter(typeof(BrickletDeviceNameConverter))]
-        public TinkerforgeHelpers.DeviceData Device { get; set; }
+        public DeviceData Device { get; set; }
 
         [Description("Specifies the period between sample event callbacks. A value of zero disables event reporting.")]
         public long Period { get; set; } = 1000;
@@ -31,7 +30,7 @@ namespace Bonsai.Tinkerforge
         {
             return source.SelectStream(connection =>
             {
-                var device = new global::Tinkerforge.BrickletAmbientLightV3(Device.UID, connection);
+                var device = new global::Tinkerforge.BrickletAmbientLightV3(Device.Uid, connection);
                 connection.Connected += (sender, e) =>
                 {
                     device.SetStatusLEDConfig((byte)StatusLed);
