+++ Bonsai.Tinkerforge/BrickletCO2V2.cs
@@ -3,7 +3,6 @@ using System.ComponentModel;
 using System.Reactive.Disposables;
 using System.Reactive.Linq;
 using Tinkerforge;
-using System.Xml.Serialization;
 
 namespace Bonsai.Tinkerforge
 {
@@ -14,7 +13,7 @@ namespace Bonsai.Tinkerforge
     {
         [Description("Device data including address UID.")]
         [TypeConverter(typeof(BrickletDeviceNameConverter))]
-        public TinkerforgeHelpers.DeviceData Device { get; set; }
+        public DeviceData Device { get; set; }
 
         [Description("Specifies the period between sample event callbacks. A value of zero disables event reporting.")]
         public long Period { get; set; } = 1000;
@@ -32,7 +31,7 @@ namespace Bonsai.Tinkerforge
         {
             return source.SelectStream(connection =>
             {
-                var device = new global::Tinkerforge.BrickletCO2V2(Device.UID, connection);
+                var device = new global::Tinkerforge.BrickletCO2V2(Device.Uid, connection);
                 connection.Connected += (sender, e) =>
                 {
                     device.SetStatusLEDConfig((byte)StatusLed);
