+++ Bonsai.Tinkerforge/BrickletDeviceNameConverter.cs
@@ -12,8 +12,7 @@ namespace Bonsai.Tinkerforge
 {
     internal class BrickletDeviceNameConverter : TypeConverter
     {
-        public Dictionary<string, TinkerforgeHelpers.DeviceData> devices;
-        //private Stopwatch stopwatch;
+        public Dictionary<string, DeviceData> devices;
 
         public override bool GetStandardValuesSupported(ITypeDescriptorContext context)
         {
@@ -44,7 +43,7 @@ namespace Bonsai.Tinkerforge
 
         public override object ConvertTo(ITypeDescriptorContext context, CultureInfo culture, object value, Type destinationType)
         {
-            var casted = value as TinkerforgeHelpers.DeviceData;
+            var casted = value as DeviceData;
             return destinationType == typeof(string) && casted != null
                 ? casted.ToString()
                 : base.ConvertTo(context, culture, value, destinationType);
@@ -70,7 +69,7 @@ namespace Bonsai.Tinkerforge
                                          .Distinct().ToList();
 
                     // For each IP connection, we search the connected devices and add it to the list - TODO test for multiple IP
-                    devices = new Dictionary<string, TinkerforgeHelpers.DeviceData>();
+                    devices = new Dictionary<string, DeviceData>();
                     foreach (ConnectionID connectionID in connectionIDs)
                     {
                         var ipcon = new IPConnection();
@@ -93,13 +92,11 @@ namespace Bonsai.Tinkerforge
             return base.GetStandardValues(context);
         }
 
-        /// <summary>
-        /// EnumerateCallback used to add devices to the device list when IPConnection.Enumerate() is called
-        /// </summary>
         private void EnumerateConnection(IPConnection sender, string uid, string connectedUid, char position,
             short[] hardwareVersion, short[] firmwareVersion, int deviceIdentifier, short enumerationType)
         {
-            TinkerforgeHelpers.DeviceData discoveredDevice = new TinkerforgeHelpers.DeviceData(uid, connectedUid, position, deviceIdentifier);
+            // Callback for IPConnection.Enumerate() that adds devices to the device list
+            DeviceData discoveredDevice = new DeviceData(uid, connectedUid, position, deviceIdentifier);
             devices.Add(discoveredDevice.ToString(), discoveredDevice);
         }
 
@@ -114,45 +111,5 @@ namespace Bonsai.Tinkerforge
                 return $"{HostName}:{Port}";
             }
         }
-
-        // Goncalo's method of doing the above, also works and probably more scaleable for more complex workflows
-        //public override StandardValuesCollection GetStandardValues(ITypeDescriptorContext context)
-        //{
-        //    // TODO - some of these need to go into a helper class for dealing with grabbing connected stuff
-        //    // Here we grab all the existing Tinkerforge IPConnection nodes as output. TODO - should give the associated UIDs for those connections
-        //    if (context != null)
-        //    {
-        //        if (stopwatch == null || stopwatch.ElapsedMilliseconds > 1000)
-        //        {
-        //            var workflowBuilder = (WorkflowBuilder)context.GetService(typeof(WorkflowBuilder));
-        //            if (workflowBuilder != null)
-        //            {
-        //                stopwatch = Stopwatch.StartNew();
-        //                Console.WriteLine("request");
-        //                // Get the distinct connection specs (host / port)
-        //                var connectionIDs = (from builder in workflowBuilder.Workflow.Descendants()
-        //                                     let createIP = ExpressionBuilder.GetWorkflowElement(builder) as CreateIPConnection
-        //                                     where createIP != null
-        //                                     select new ConnectionID { HostName = createIP.HostName, Port = createIP.Port })
-        //                                 .Distinct().ToList();
-
-        //                // Test uid
-        //                devices = new List<string>();
-        //                var ipcon = new IPConnection();
-        //                ipcon.Connect("localhost", 4223);
-
-        //                ipcon.EnumerateCallback += EnumerateConnection;
-        //                ipcon.Enumerate();
-        //                Thread.Sleep(100);
-
-        //                ipcon.Disconnect();
-        //            }
-        //        }
-
-        //        return new StandardValuesCollection(devices);
-        //    }
-
-        //    return base.GetStandardValues(context);
-        //}
     }
 }
